Renamed package from testnaming to naming
