+++ src/main/java/junitparams/naming/TestCaseName.java
@@ -1,4 +1,4 @@
-package junitparams.testnaming;
+package junitparams.naming;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
